weighted repeat penalty

+++ generator/gpt2/src/sample.py
@@ -6,8 +6,8 @@ from generator.gpt2.src import model
 def penalize_used(logits, output):
     # output has shape (1, len) and type int32
     n_vocab = logits.shape[1]
-    N = tf.minimum(180, output.shape[0])  # lookback
-    weights = tf.range(1, N + 1, dtype=tf.float32) / float(N)
+    N = tf.to_float(tf.minimum(180, output.shape[0]))  # lookback
+    weights = tf.range(1, N + 1, dtype=tf.float32) / N
     counts = tf.math.bincount(output[0, -N:], weights=weights,
                               minlength=n_vocab, dtype=tf.float32)
     counts = tf.expand_dims(counts, 0)
